requirements.txt
assertpy==1.1
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.1
python-decouple==3.7
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
